split DefaultGamesLocator to extract all GameSpecSource classes
