+++ gazeplay/src/main/java/net/gazeplay/games/bubbles/ColoredBubblesGameSpecSource.java
@@ -1,0 +1,15 @@
+package net.gazeplay.games.bubbles;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class ColoredBubblesGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("ColoredBubbles", "data/Thumbnails/bubblecolor.png", GameCategories.Category.SELECTION),
+				new ColoredBubblesGameLauncher());
+	}
+}
