+++ gazeplay/src/main/java/net/gazeplay/games/drawonvideo/VideoPlayerGameSpecSource.java
@@ -1,0 +1,15 @@
+package net.gazeplay.games.drawonvideo;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class VideoPlayerGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("VideoPlayer", "data/Thumbnails/youtube.png", GameCategories.Category.ACTION_REACTION),
+				new VideoPlayerGameVariantGenerator(), new VideoPlayerGameLauncher());
+	}
+}
