+++ gazeplay/src/main/java/net/gazeplay/games/horses/HorsesSimplifiedGameSpecSource.java
@@ -1,0 +1,16 @@
+package net.gazeplay.games.horses;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.horses.HorsesSimplifiedGameLauncher;
+import net.gazeplay.games.horses.HorsesSimplifiedGameVariantGenerator;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class HorsesSimplifiedGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(new GameSummary("Horses Simplified", "data/Thumbnails/horsesSimplified.png",
+				GameCategories.Category.ACTION_REACTION), new HorsesSimplifiedGameVariantGenerator(), new HorsesSimplifiedGameLauncher());
+	}
+}
