+++ gazeplay/src/main/java/net/gazeplay/games/math101/Math101GameSpecSource.java
@@ -1,0 +1,16 @@
+package net.gazeplay.games.math101;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.math101.Math101GameLauncher;
+import net.gazeplay.games.math101.Math101GameVariantGenerator;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class Math101GameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(new GameSummary("Math101", "data/Thumbnails/math101.png",
+				GameCategories.Category.LOGIC, null, "MathDescAdd"), new Math101GameVariantGenerator(), new Math101GameLauncher());
+	}
+}
