+++ gazeplay/src/main/java/net/gazeplay/games/memory/MemoryGameSpecSource.java
@@ -1,0 +1,17 @@
+package net.gazeplay.games.memory;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.memory.MemoryGameLauncher;
+import net.gazeplay.games.memory.MemoryGameVariantGenerator;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class MemoryGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("Memory", "data/Thumbnails/memory.png", GameCategories.Category.MEMORIZATION),
+				new MemoryGameVariantGenerator(), new MemoryGameLauncher());
+	}
+}
