+++ gazeplay/src/main/java/net/gazeplay/games/memory/MemoryLettersGameSpecSource.java
@@ -1,0 +1,16 @@
+package net.gazeplay.games.memory;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.memory.MemoryLettersGameLauncher;
+import net.gazeplay.games.memory.MemoryLettersGameVariantGenerator;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class MemoryLettersGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(new GameSummary("MemoryLetters", "data/Thumbnails/memory-letter.png",
+				GameCategories.Category.MEMORIZATION), new MemoryLettersGameVariantGenerator(), new MemoryLettersGameLauncher());
+	}
+}
