+++ gazeplay/src/main/java/net/gazeplay/games/memory/MemoryNumbersGameSpecSource.java
@@ -1,0 +1,14 @@
+package net.gazeplay.games.memory;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class MemoryNumbersGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(new GameSummary("MemoryNumbers", "data/Thumbnails/memory-numbers.png",
+				GameCategories.Category.MEMORIZATION), new MemoryNumbersGameVariantGenerator(), new MemoryNumbersGameLauncher());
+	}
+}
