+++ gazeplay/src/main/java/net/gazeplay/games/memory/OpenMemoryGameSpecSource.java
@@ -1,0 +1,16 @@
+package net.gazeplay.games.memory;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.memory.OpenMemoryGameLauncher;
+import net.gazeplay.games.memory.OpenMemoryGameVariantGenerator;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class OpenMemoryGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(new GameSummary("OpenMemory", "data/Thumbnails/openMemory.png",
+				GameCategories.Category.ACTION_REACTION), new OpenMemoryGameVariantGenerator(), new OpenMemoryGameLauncher());
+	}
+}
