+++ gazeplay/src/main/java/net/gazeplay/games/soundsoflife/FarmGameSpecSource.java
@@ -1,0 +1,15 @@
+package net.gazeplay.games.soundsoflife;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class FarmGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("Farm", "data/Thumbnails/farm.png", GameCategories.Category.ACTION_REACTION),
+				new FarmGameLauncher());
+	}
+}
