+++ gazeplay/src/main/java/net/gazeplay/games/soundsoflife/JungleGameSpecSource.java
@@ -1,0 +1,15 @@
+package net.gazeplay.games.soundsoflife;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class JungleGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("Jungle", "data/Thumbnails/jungle.png", GameCategories.Category.ACTION_REACTION),
+				new JungleGameLauncher());
+	}
+}
