+++ gazeplay/src/main/java/net/gazeplay/games/soundsoflife/SavanaGameSpecSource.java
@@ -1,0 +1,15 @@
+package net.gazeplay.games.soundsoflife;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSummary;
+import net.gazeplay.gameslocator.GameSpecSource;
+
+public class SavanaGameSpecSource implements GameSpecSource {
+	@Override
+	public GameSpec getGameSpec() {
+		return new GameSpec(
+				new GameSummary("Savanna", "data/Thumbnails/savana.png", GameCategories.Category.ACTION_REACTION),
+				new SavannaGameLauncher());
+	}
+}
